Add script to generate all *-base.h's.
In order to resolve b/31800672, we added a target to hidl,
-Lexport-headers, which creates a C compatible header file that targets
can link against to get access to hidl constants without having to link
against the hidl code directly. However, these files are (for some
reason) not generated at build time. Until they are, this file can be
used to update and keep track of them.

Bug: 32883144
Test: ran script
Change-Id: Ib07433f98c4152ca34b48789dcb117eb262241cb