contexthub: Explicitly pass nanoapp binary header data
Parse the nanoapp binary header in the framework and pass the relevant
data to the HAL explicitly in the NanoAppBinary structure. Note that not
all fields from the header are carried forward into NanoAppBinary, as
they are either implied by the HAL API (version, magic), or inherent in
the implementation-specific data (hwHubType), or unused and will only be
added via a future HAL version bump (second reserved field).

Also, convert the flags field in struct MemRange to a bitmask, which is
the proper representation for its use.

Bug: 34182159
Test: run hello world context hub GTS test
Change-Id: I13da9883b851e978d760548ec4265094ce81510a