Post-O. Finalize Vehicle HAL support for OBD2 diagnostics.
Add OBD2_DTC_INFO property to VHAL. This property returns a list of timestamps for which freeze frames are available to be read.
Make OBD2_FREEZE_FRAME indexed on timestamp, such that one has to pass a timestamp to retrieve the freeze frame at that timestamp.
Add OBD2_DTC_CLEAR to delete some or all of the freeze frames stored.

Test: build
Change-Id: I30344ffd6e7527f076b3382c32f5507973c9985b