Use libstagefright_foundation_vendor
DRM HAL is a library in vendor. When BOARD_VNDK_VERSION is turned
on, any lib in vendor partition can only link against to other vendor
libs or libs marked as vendor_available:true. This lib is using
libstagefright_foundation and the lib is not available to vendors.
Instead, libstagefright_foundation_vendor is provided as a
vendor-variant of the foundation library. Thus switching to the vendor
variant.

Bug: 37462569
Test: m android.hardware.drm@1.0-impl
Change-Id: Iaee9886f9ba3643e1c19928690bad13c26ac29d7