bug fix: handle portReconfiguration in testEOS()
When CSD and sync frame are sent together, component parses the
CSD and allocates buffers internal to the codec and decodes the frame
sent and port reconfiguration event is sent to the client. Now if this
is the only frame in the sequence, the client reaches testEOS(). But
port reconfiguration is not handled here. This is corrected.

bug: 62930419

Change-Id: Ic0c7625e3c2b415330f37734663f0cf7689e8734