configstore: Allow syscalls needed by crash_dump
Information such as tombstones and callstack are important when
debugging why a process crashed. configstore's seccomp filter
and selinux policy are currently overly strict and are blocking
the collection of crash data.

In order to keep configstore's sandbox as strict as possible, use
Crash_dump's fallback mechanism to collect crash data. This uses a
local socket to send crash data to tombstoned.

Bug: 64768925
Test: killall -ABRT android.hardware.configstore@1.1-service
    Verify that configstore callstack gets dumped to logcat.
    Verify that a crash tombstone gets recorded to /data/tombstones
Test: proto_fuzzer runs without crashing

Change-Id: I3c3e13d6aa1c1e2dda2a619dfa815375ee7ebff6