Fix VTS test which turns off radio.
Since the purpose of most of the test cases in current VTS is to check
if proper errors are returned and there is no crash seen in vendor code,
updating setRadioPower test case to turn on the radio instead of
turning off. We want to avoid test cases which turn off
radio or leads to modem shut down as those test cases affect other tests.

Test: VTS
Change-Id: I4fb9f18884f7ef21162015a0032c4431444f7025
Bug: 65230472