tetheroffload VTS: fix stopOffload test conditions
The tetheroffload HAL is somewhat...and over- and under-specified.
A not unreasonable interpretation is that stopOffload() doesn't have
return success unless offload was actually started (via a call to
setUpstreamParameters()), and that after initOffload() and before
setUpstreamParameters() not actual "offload" has been engaged.

Precision in this matter is not required for the test case:

    OffloadControlHidlTestBase.AdditionalInitsWithoutStopReturnFalse

But for the test case:

    OffloadControlHidlTestBase.AdditionalStopsWithInitReturnFalse

we want to ensure the "matching" stopOffload call succeeds. For this
test we add in a call to setUpstreamParameters() for good measure.

Test: as follows
    prompt$ make vts -j30 BUILD_GOOGLE_VTS=true && \
            vts-tradefed run commandAndExit vts \
                --skip-all-system-status-check \
                --primary-abi-only \
                --skip-preconditions \
                --module VtsHalTetherOffloadControlV1_0Target \
                -l DEBUG

Observed:
    10-16 19:17:17 I/ResultReporter: Invocation finished in 1m 2s. PASSED: 38, FAILED: 0, MODULES: 1 of 1

Bug: 65270149
Bug: 65612227
Bug: 65612332
Change-Id: I924d41f5a20f07707e3d6991cb59d9c6b2b02339